Fix inconsistent personalisation with route guards
WHen a user tries to access a route-guard protected resource and that
user does not have access to that resource, the personalisation service
might still set the color for that resource.

The personalisation service listened to the NavigationStart event which
is launched even if the router cannot go to the selected path.

Changing it to the NavigationEnd event should fix it.